refactor(profile): align metric ownership and naming
